[CLOTA] Add new build format to events
BZ: 23284

Before, not enough information were reported to distinguish
if a build was official or not.

In improving the number of information reported, we make a
unique signature which allow to know if a build was modified.
Reported elements composing the build signature (properties) :
 - ro.build.version.incremental
 - ro.build.fingerprint
 - sys.kernel.version
 - ro.build.user@ro.build.host
 - gsm.version.baseband
 - sys.ifwi.version
 - sys.ia32.version
 - sys.scu.version
 - sys.punit.version
 - sys.valhooks.version

Fix an upload logs notification bad condition too.

Change-Id: I0e07dc3395b3cf99c77e04b2f3ad7043736e9371
Signed-off-by: Rocher, JeremyX <jeremyx.rocher@intel.com>
Reviewed-on: http://android.intel.com:8080/37291
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>