[CRASHLOG] Manage log report by exclusion type
BZ: 28162

Reason for the change : Need to change the algorithm in order to exclude a few
type from uploading their logs

Changes done :
Now, log type to upload is managed by exception.
Preferences screen now store log type to exclude (instead of log type to upload)
By default, all types are upload expected AP core DUMP.
The selection of log to upload is calculated with a query using a "NOT IN" clause.
This clause is constructed with type excluded.

Change-Id: Ie3f0c24cc254ec99ac9ef39efcc313311f156f1e
Signed-off-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-on: http://android.intel.com:8080/42400
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Rocher, JeremyX <jeremyx.rocher@intel.com>
Tested-by: Zhang, Dongxing <dongxing.zhang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>