[CLOTA] Display a notification when a critical crash occured
BZ: 32392

When a critical crash occured there is no way to be notified of the event
unless the crash report management is "ask for upload". This may be a problem
when the device reboot due to a critical crash and the user doesn't see it when
that happens.

To solve this problem an android notification is now displayed when a crash
occurs.The critical crash type list is set in a private Android preference.
The parameter "notified" has been added to the table "events" of the database to
know if an event has been notified or not.
An event is not notified until the user doesn't remove the notification.
The table "critical_events_type" has been added to the database to know if a
crash type is critical or not.

Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>

Change-Id: I900272a0e0d5f7d20bf4ad7630d983122dbaa081
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/43905
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Zhang, Dongxing <dongxing.zhang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>