[CLOTA] java.lang.NullPointerException happens in com.intel.crashreport
BZ: 31111

Javacrash happened in differents cases (monkey test, crashreport test,
manual CrashReport.apk installation) due to a java.lang.NullPointerException in
com.intel.crashreport. This is due to a bad initialization of the CrashReport
Service.
Indeed, when the StartServiceActivity tries to bind to the
CrashReportService it uses the bindService function which creates the
CrashReportService. In some cases the bind could be done before the end of the
service's creation. In those cases the use of the service generates
a NullPointerException.

To solve this problem a check has been added to the getService() function used
in the bindService operation. The goal is to check that the service exists.
Now the function wait until the end of the service's creation before to give it
to the bind service function.

Change-Id: Ifea362d54dcd0fc449a82c0d187a6563df968499
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/47025
Reviewed-by: Rocher, JeremyX <jeremyx.rocher@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Zhang, Dongxing <dongxing.zhang@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>