[CLOTA} Notify System_server's TOMBSTONE
BZ: 33705

The crashes's notification's system didn't display the server_system tombstone.

To solve this problem a new implementation of the system was done.
A new table critical_events was created to store the signature (type,data0...5)
of the critical crashes (as system_server tombstone).
The table critical_events_type was renamed in events_type to describe better
its function.
Each time an events occurs, the system checks if its type is critical or not
and if it matches with a critical crash signature (it's done in processEvent()
in CrashReportService). If an event is critical, the system informs the user
by an android notification.
The critical types and critical crashes are added in the database at
the creation of CrashReport.

Change-Id: I1a67db1e2287eddc28d2afe90227c295f583e1f9
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/48445
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>