[CLOTA] Block manual upload when crash report disable property is set
BZ: 36000

The crash report disable property allows the user to stop the crahes report
whatever report management he chooses. The problem was that the manual crash
report wasn't blocked.

To solve this problem a test is set in the askForUpload() function in order to
check if the crash report disable property is set or not.

Change-Id: I596f53de1ca9264a635b1a7909785531418a6b71
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/48796
Reviewed-by: Rocher, JeremyX <jeremyx.rocher@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>