[CLOTA] Improve Unknown Event management reported to crashreport
BZ: 37992

With the apparition of the new event type STATS a problem occurs on
crashreport. When the user wants to upload STATS events a
java.lang.NullPointerException occured. It's due to a wrong management of
the unknown type events reporting.
Indeed, in case of unknown events type the event is created and filled with a
null field for the date and a blank field for the other parameters. When the
application tries to add the event in the database the date conversion doesn't
work because the date is null and it throws the nullPointerException.

To solve the problem a test is done in the convertDateForDb() function
in EventDB. This test checks that the Date is not null before to use it.
If the date is null the function returns -1.
Another test is done in addEvent() function in EventDB. This function checks
that the name and the Date is known before to add an Event in the Database.

Change-Id: I68beefc829cc1c3b17efd8f640a1e0bb34c9e2fa
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/50248
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>