[PORT FROM R3][CLOTA] Change alarm type from RTC_WAKEUP to RTC
BZ:39472

When upload fails, an alarm is set. RTC_WAKEUP alarm was used
and wakes up the device when it was in deep sleep (S3).
This behavior results in power overconsumption in IDLE UCs
which is not desired.

Change the alarm type to RTC will note wakes up the device when
it is in deep sleep state.

Change-Id: I2c9a54c9eb17388004bf6c67a363592219aa9dc4
Signed-off-by: Jeremy Rocher <jeremyx.rocher@intel.com>
Reviewed-on: http://android.intel.com:8080/51807
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>