[CLOTA] Rework on the UI
BZ: 39196

In order to get new features on CrashReport and so new frames(Activities)
a rework is welcome. This rework will also display the defect reporting and
stats upload.

In this rework, the settings and actions are separated. One Activity displays
the settings menu and another displays the actions menu.
The first one displays all the settings available for CrashReport. The second
one displays a list of buttons. Each button indicates an action that can be
lunched on CrashReport. For the moment the action available is "Report Crash
Event" but new actions will be available soon (create a bugzilla,logs
configuration).

The dialog has been changed and replaced by DialogFragment in order to be
compliant with the new Android version.

Change-Id: Ia90d8f7f917f36663c9c50da7337be74ae059af6
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/53090
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Rocher, JeremyX <jeremyx.rocher@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>