[PORT FROM R4][CLOTA] Resolve DialogFragment crashes in CrashReport
BZ: 45680

Some issues were founded in CrashReport due to the use of DialogFragment
instead of Dialog.
Indeed DialogFragment aren't managed like the Dialog and need to be more
vigilant with them and their lifecycle.
Contrary to Dialog, DialogFragment are not attached to the activity in which
they are created. For this reason, when the user presses the Back button in
order to cancel the StartServiceActivity, the AskDialog which displays the
CrashReportManagment (if the display is needed) is not killed before the
Activity.

One bad sequence which causes an issue is:
The user launches the StartServiceActivity and presses immediatly the Back
button.The activity is immediatly stopped but the AskDialog is only
displayed after the stop of the Activity, and this is not compliant with
Android policies. For this reason, a crash occurs.
The second bad sequence which causes an other issue is:
The user launches the StartServiceActivity and presses immediatly the Back
button.The Activity is immediatly stopped and the AskDialog is destroyed
after.In its destruction the AskDialog calls OnCancel() and sends a message to
the CrashReport service, but the service is destroyed during the stop of
the Activity and a crash occurs(Null Pointer).

To solve the problems a boolean was created in order to be aware of the stop of
StartServiceActivity. When the display of CrashReport Managment is asked, a
check on this boolean is done to display or not the dialog.
Furthermore, when a press on the Back button is done in StartServiceActivity,
a check is done to close the CrashReport Managment if it exists.

Change-Id: I8914c9964f512309ba5104c42b4ed842e20c0b9e
Org-Change-Id: I8a2086c186c193c722d9cc2a557149531bbc356a
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/59849
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>