[CLOTA] Add manual aplogs upload
BZ: 50990

There are more and more reports from the field and the upper management of big
battery drain. In all those cases, it is pretty difficult to get access to the
aplog. In order to solve this problem, it could be a good thing to add the
possibility to upload aplogs directly from the device in pushing a dedicated
button on CrashReport.

A new button "upload aplog" was added to the home.xml layout in order to add
in the CrashReport UI the possibility to upload aplogs.
When this button is pressed, a new trigger was created. This trigger is
aplog_trigger file and it is created in /data/logs/aplogs repository.
When this trigger is received by crashlogd new APLOG events were created and
correpond to the aplogs to send.
When a data connection is available and CrashReport is in the upload state,
CrashReport send APLOG events like the other events.
When CrashReport sends data associated with an APLOG event a ProgressBar is
displayed in StartServiceActivity.This progress bar corresponds to the data
upload progression. When a APLOG data upload is finished, the progress bar
disappears.
New messages showProgressbarView,hideProgressbarView and uploadProgressbarView
were added in StartServiceActivity in order to control the display and
progress value of the progress bar.
Some modifications were made in Connector.java in order to compute the data
upload progress.

Update : manage crashlogd missing + alertdialog for upload

Conflicts:

	crashreport/res/layout/home.xml
	crashreport/res/values/strings.xml
	crashreport/src/com/intel/crashreport/CrashReportHome.java

	modified:   crashreport/res/layout/home.xml
	modified:   crashreport/res/values/strings.xml
	modified:   crashreport/src/com/intel/crashreport/CrashReportHome.java
Update : resolve conflict with logconfig
Update2 : minor change for progressbar display
Update3 : migrate from JB to Main
Update4 : fix progress bar parameter + filesize issue

Change-Id: I2fd0a9eb6a3fb23f1026dc2a9190f6a4035fb77b
Signed-off-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-on: http://android.intel.com:8080/61688
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>