[CLOTA] Support a delay if necessary before uploading crashlogs
BZ: 53700

With the add of dumpstate and dropbox logs in the crashlogs directory, it could
take a long time before a crash be ready to upload. This is due to
the fact that dumpstate and dropbox take a long time to process. It's became
necessary to add a delay in order to postpone the upload of a crash.

To solve this problem, modifications were done in CrashReport. When a crash
occurs crashlogd starts dumpstate and dropbox. CrashReport is immediatly
notifyed of the crash's presence, but the crash is tag as not "ready for
upload" and so CrashReport doesn't try to upload it. A timer starts and it will
end after an amount of time fixed at 60sec. A crash is taged as "ready for
upload" when the timer ended or when dumpstate and dropbox finished, depending
on which finished first. When a crash is ready for upload, CrashReport is
immediatly notified and try to report the crash.

A new field "dataReady" was added in the events table.This field represents the
state of the crash logs. 1 if the logs are ready to upload and else 0.
New functions (updateEventDataReady() and eventDataAreReady()) were added in
EventDB.java to update the dataReady field and to check if crash is ready for
upload or not.
A new class NotifyCrashTask was added in CrashReportService.java. This is the
timer task that will be excuted at the end of the crash upload postpone delay
time. It sets the dataReady field to 1 if it's not already done and starts
CrashReportService.

Update : change timeout to 120 seconds
  added data_ready to crashinfo

Change-Id: I1bba919839510e3e4db0200b6402c827eb41d061
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/64185
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>