[CLOTA] Fix a Javacrash in CrashReport due to a bug with the upload progress bar
BZ: 52294

A bug was found in CrashReport. This bug is due to the update of the upload
progress bar.
Indeed, when the application try to update the upload progress
bar it sends an intent to StartServiceActivity in order to modify the progress
current value of the progress bar. The problem happens when CrashReportService
and StartServiceActivity are not linked.

To fix the problem a check is done to be sure that service and activity are
bounded before the broadcast of the updateProgressBar intent.

Change-Id: Ic6c2993fe729fdb96159a7a8782c9a44541e44c0
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/64543
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>