[CLOTA] Crash time shown on Crash Tool log Analyzer is no correct
BZ: 53405

Crash time displayed on CrashTool log analyser is not the same that Crash time
written in the crashfile. This is due to the fact that CrashTool timezone and
device timezone may not be the same.

To solve this problem the date read in history_event is parsed as an GMT time
and not as a Local Device Time. In this way there is no time conversion.
Then, to be sure that CrashTool displays the correct date, a CrashTool Timezone
date is created with the original date.

Change-Id: I38362715da9b7525ad8c0b2a99e076bcc68935e2
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/65464
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>