[CLOTA] Improve Report a Bug
BZ: 59374

There was a bad management of return lines in the "report a bug" feature of
PhoneDoctor.
This is due to a use of a wrong Java function. Indeed, replaceAll() was used to
replace all '\n' by '\\n' in the bz description file. It's better to use
the replace() function instead.

Moreover, some code was added to refresh the Gallery view when the user change
the checkBox value.

Change-Id: I130e8c70712b79feb0021ed7dd4b2dc3aa03de6d
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/67915
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>