[CLOTA] Add database right access for crashinfo
BZ: 59861

In user debug_build, CrashInfo is not root and doesn't have access to the
database. In this case all functions which communicate with the database don't
work.

To solve the problem it's necessary to give the good access right to the
database for crashinfo.
This is done in EventDB.java. The DatabaseManagerHelper doesn't extend
SQLiteOpenHelperManager anymore. Indeed, the open function of SQLiteOpenHelper
changes the access right of the database each time it's called.
DatabaseManagerHelper manages the update/create/open/close operation on the
database and changes the access right if it's necessary.
Moreover, at each reboot of PhoneDoctor, the database access right are changed.

Change-Id: Ia2c82fef9fe3dca8f622b37477251c82546d0133
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/68388
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>