PhoneDoctor: Add Kratos dedicated receiver
BZ: 62593

In order to catch events from other applications, a new package is
added : com.intel.crashreport.receivers. This package will receive
classes which extends BroadcastReceiver. They must be added to the
AndroidManifest to be effective. Once intent are received, an event is
generated and sent to crashtool.

First receiver is dedicated to Kratos to receive its
intel.intent.action.kratos.SESSION_INFO intent. It will generate an
INFO event of type KRATOS with "Session upload" as DATA0 and the link
as DATA1.

Change-Id: Ideefb6cce7a8414c4b3702a36b4cb8285d945cfe
Signed-off-by: Jeremy Rocher <jeremyx.rocher@intel.com>
Reviewed-on: http://android.intel.com:8080/70328
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>