[CLOTA] Improve picture management in PhoneDoctor Bugzilla Activity
BZ: 60186

When the user try to open a picture with PhoneDoctor and when this picture
is not stored in the media storage place, PhoneDoctor crashes.
This is due to the fact that a ContentResolver is used to get the pathname of
the picture selected. This ContentResover is like a SQL request and executes a
request on the phone system files. This request uses columns that only exist
for picture stored on the disk and not for picture on the web.

To solve the problem the function getColumnIndexOrThrow was replaced by
getColumnIndex. Indeed the first one throws an exception if the column doesn't
exist.

Change-Id: I831c716966a071275009801b93f7f49af84e8809
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/69212
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Rocher, JeremyX <jeremyx.rocher@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>