[PORT FROM MAIN][CLOTA] Check new Event during upload
BZ: 43452

While PhoneDoctor is uploading events and logs, it can't check the presence
of new events. So all events which occur during the upload time are not catched
by PhoneDoctor and have to wait the end of the upload.
Moreover events are not catched immediatly after the events upload, they have
to wait a new event notification in order to launch the history_event parse.
In this way, it's possible to wait an huge amount of time before to detect new
event.

To solve the problem it's important to create a new Thread in the
NotificationReceiver in order to add each new event in the database if
CrashReportService is already launched.
CrashReportService has been modified in order to check the presence of a new
Event after events upload and between two logs upload. Logs upload could take
a lot of time (especially over 3G) and during this time some events could
appear.

Change-Id: I0d909055151b3a8d97ec6c83d2fbccc0555e31af
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/69770
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Pince, Jean-christophe <jean-christophe.pince@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/78513
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>