[CLOTA] Add a lock logconfig system
BZ: 60739 63156

Some logconfigs need to be always enabled, and users want to never modify them.
Monkey test or human mistakes can disable those logconfigs and it becomes
necessary to avoid manual changes on those configs.
To solve the problem a new type of config was added: the lock logconfig.
The lock logconfig are enabled at each reboot and the switch button associated
with is disabled. In this way users can't interact with it and can't disable
the config.

Change-Id: Ib591d05f089a0757075e57ceb2ccb6a4faaddd8e
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/71124
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>