CLOTA: "Report a bz" improvement
BZ: 63743 59460

In "Report a bug" PhoneDoctor feature there are few things that can be
improved.
In the BZ creation page, the description field is limited to 5 lines with no
scroll bar. When the user wants to type a text bigger than 5 lines he can't
read the first lines of his description and it's annoying.
Still in this same page, when the user wants to switch application and launch
another application the content of the BZ is not saved and user has to re-write
all the content when he comes back to the BZ creation page.
In the summary page there is no scroll bar and it's annoying when the
BZ description text is big, user can't see all the content of his BZ.
PhoneDoctor catches all intents related to picture interaction and when
the user wants to attach a picture to an MMS in the CHOOSE APP popup PhoneDoctor
is present.It could be confused for the user.

Layout were modified to solve those problems.
In activity_bugzilla_main.xml, the limited number lines of the description
field was removed in order to display all the text.
In bugzilla_summary.xml, a scroll bar was added to scroll over all the content
of the created BZ.

Some modifications were made in BugzillaMainActivity in order to save BZ
on the activity pause.

Change-Id: Ie29fcbcf04170a4de0728a89f98c199dfb9d5421
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/71840
Reviewed-by: Pince, Jean-christophe <jean-christophe.pince@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>