[PORT FROM MAIN]CLOTA: Turn off auto-correct for edittext widgets
BZ: 64865

The auto-correct system is present for the editText in the BZ creation
activity. It makes auto-correction on the text typed by the user. In title
and summary field, user usually uses shorthand/abbreviate/etc and
the auto-correction system corrects them with wrong words.It's necessary to
disable the auto-correct system.

To solve this problem the auto-correction was removed for the title and summary
field. Moreover suggestions system was removed too.

Change-Id: I6bf035104f0e268f6fe175484aa528ea68fa81a8
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/73779
Reviewed-by: lab_cactus <lab_cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Reviewed-on: http://android.intel.com:8080/78510
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>