[PORT FROM MAIN]CLOTA: Multiple screenshots selection
BZ: 64301

Some users are confused with the screenshot selection in the BZ creation
on PhoneDoctor. They can't make the distinction between which screenshot is
selected and which are not. Moreover the possibility to attach several
screenshots to a BZ could be a good thing.

The aspect of a screenshot selected was modified. Now when a screenshot
is selected it's surrounded by a blue rectangle. screenshot_item.xml was
modified to change the padding of the ImageView in order to display the blue
rectangle.
The behaviour of the activity was changed too. When a user click on a
screenshot, the screenshot becomes selected if it wasn't and unselected if it
was. A HashMap was added in ScreenshotAdapter.java in order to store
selected screenshots. It represents the list of each screenshot present in the
screenshot directory  with a boolean to inform if the screenshot is selected.
The bz_trigger created by PhoneDoctor when the report button is pushed was
modified. It contains now a SCREENSHOT= line for each screenshot attached to
the BZ.
BugStorage.java was modified too to store the list of screenshots selected
in a HashSet preference.

Change-Id: I5318533c86855cb7874b46e4ff2a97c14a4a9796
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/76560
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-on: http://android.intel.com:8080/80322
Tested-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>