CLOTA: Javacrash on PhoneDoctor
BZ: 75146

Sometimes a Javacrash occurs on PhoneDoctor when the user wants to cancel or
approve the report managment dialog in the events viewer activity.
This is due to the fact that the CrashReportService is died or unbound.

To solve this problem the NotificationReceiver was changed. When an intent is
received,if CrashReportService is not started the application launchs it. Else,
the checkEvents() function of CrashReport is launched to check the presence of
new events. This modification avoid to call the service destruction instruction
that is present in onStartCommand() in CrashReportService. Indeed, in case of
an already started service, the method kills it.
Moreover, some modifications was done on StartServiceActivity to be sure that
the reportManagment dialog is destroyed when the service unbind.

Change-Id: Ie6caf76f29b3e2d64a4a42a54cb8fe2e44c54338
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/86852
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Guerard, Christophe <christophe.guerard@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>