PhoneDoctor & crashinfo : adding event field
BZ: 73818

In order to add field to phone doctor and spread it
through crashtool chain

Phone doctor is updated to manage new field pdstatus
crashtool beans are updated to manage :
- row id
- origin
- ssn
- pdStatus

Crashinfo is also updated to display pdstatus
in full mode

PDStatus field is composed by a 50 characters String which represents
the informations below:
Check if an modem coredump is linked with the MPanic or not.
Check if the EventId in the database is the same that the EventId in the
crashfile.
Check if the dropbox is full or not.
Check if crashlogger daemon is still running or not.
Check if apklogfs is running or not.
Check if there is enough space available on the /logs partition.
Check if the PhoneDoctor database is available.
Check if a sdcard is present on the device and if it is available.
Test the value of the persist.core.enabled property.
Check if the history_event file is not corrupted.
Check if logs have to be upload or not.
Check if logs have to be upload only over a Wifi connection.
Number of uptime since the last SWUPDATE event.
Check if the device is connected to an Wifi network or not.
Check if the device is connected to a 3G network.

The PDStatus class computes the PDStatus of an event.

Signed-off-by: Benoit, NicolasX <nicolasx.benoit@intel.com>

Change-Id: I8379bc936d883ff75bfe6b160d3956fa8be9efc0
Reviewed-on: http://android.intel.com:8080/90585
Reviewed-by: Vidoine, Charles EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>