CLOTA: Crash rain not detected
BZ: 86124

Crashes of a rain of TOMBSTONE were not detected as part of the rain.
The direct consequence is that some crashes were inserted in the PhoneDoctor
database instead of be blacklisted.
Moreover, when crashes occur very fast, PhoneDoctor can duplicate some crashes.
This is due to the fact that multiple threads can be running in the same time.
Those threads call the checkEvents() method from CrashReport class which is in
charge of the history_event parsing.

To solve those problems 2 modifications were made:
- The check made on the last event crash date of a rain and the new event date
to test was replaced by a less or equal instead of a less.Indeed, some crashes
could have the same date and the second event tested is rejected from the rain.
- A synchronized tag was added and surrounds the history_event line parsing.

Change-Id: I235f6a0eaed5eb165d33eb2240db700433b6d282
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/91694
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>