PhoneDoctor: Rain of crashes enhancement
BZ: 87510

Some rain of crashes are not detected for tombstone events since the crashes
signature is not complete.  To solve this problem it had been decided to
implement following evolutions:
- update of tombstone events parsing to detect the fault address in
  tombstone_xx file and set DATA3 field with this value
- update of rain of crashes algorithm in order to detect rain of tombstones
  with a new signature based on DATA0,1 and 3 fields from tombstone crashes.

Moreover, the cleaning of crashlogxx directories linked to tombstones events
included in a rain event shall be implemented in order to make room for others
more relevant crashes.

Change-Id: Icff1ce157e8e8d2a5ce4abfb7152f8ffb81b18dc
Signed-off-by: Thiry, JeanX <jeanx.thiry@intel.com>
Reviewed-on: http://android.intel.com:8080/92139
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>