PhoneDoctor: Fix default settings issues
BZ: 90725

Default values of settings where stored in different places for :
 - report management
 - event data files report
 - excluded crash data types

This is corrected by using a single source of default settings in xml
resource files.

Application version is updated to force store of default values. This
minor version should be updated at each release to ensure update of
settings and to ease error identification as this version name is now
reported with crash.

Change-Id: I2b038eeb66bc84a53f05fd1764595dcc3985f583
Signed-off-by: Jeremy Rocher <jeremy.rocher@intel.com>
Reviewed-on: http://android.intel.com:8080/94980
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>